+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/BlockNoteEditor.tsx
@@ -419,6 +419,8 @@ export const BlockNoteReader = ({
 
   useHeadings(editor);
 
+  useScrollToBlockAnchor();
+
   return (
     <Box>
       <DocsEditorStyle />